added Valve object, confirmed OOP works

+++ src/Valve.cpp
@@ -2,13 +2,13 @@
  * File              : Valve.cpp
  * Author            : Wayne Yeo [fishnsotong] <wwzyeo@gmail.com>
  * Date              : 2020-03-26T06:06:28+0800
- * Last Modified Date: 2020-03-26T06:30:45+0800
+ * Last Modified Date: 2020-03-27T14:24:54+0800
  * Last Modified By  : Wayne Yeo [fishnsotong] <wwzyeo@gmail.com>
  */
 
 #include <Valve.h>
 
-Valve::Valve(int pin, int openPos, int closePos) {
+Valve::Valve(byte pin, int openPos, int closePos) {
   this -> pin = pin;    // what is the 'this' pointer supposed to do?
   position = 0;
   init();
